No-op events unknown on StoryLive
Since StoryLive is responsible for rendering components, any events from
the components are sent to StoryLive. If handle_event is called for an
unknown event, a FunctionClauseError will be raised, crashing the
LiveView. This can be particularly problematic for any components that
trigger an event on render.

This commit adds a catch all clause to handle_event/3